requirements.txt
streamlit==1.30.0
playsound==1.3.0
SpeechRecognition==3.10.1
openai==1.9.0
python-dotenv==1.0.1
pydub==0.25.1
